requirements.txt
streamlit==1.31.0
pandas==2.1.0
numpy==1.26.0
scikit-learn==1.3.0
plotly==5.18.0
requests==2.31.0
pillow==10.0.0
matplotlib==3.8.0
joblib==1.3.2 